angle code kind of added
I added a bunch of stuff that is not being used when you run it. The angle code is kind of just there as possible ways of dealing with the angles. What would work best is if we could find a way to rotate about a point and then save the endpoint of the new branch (if we draw something and rotate it, it's endpoint coordinates will be messed up). I'm not super sure how to do this though, we need a way of doing all these transformations either as we make the points. Or we could make one point, do a transformation and store it's endpoint.