We can draw a leaf!!!
But the leaf can only be drawn in the xy plane. Adding a z component is what causes the two parts of the leaf to not match up. Note that my start and end point are randomly chosen. Perhaps we should define the end point based on the start point somehow? Like, find an end point based on where the z should end up?